Eliminate .toDouble from Firm.scala — domain-typed arithmetic (#77)
* Eliminate .toDouble from Firm.scala — use domain types

- Add PLN * Ratio → PLN and Ratio * PLN → PLN operators
- Add Ordering[PLN], Ordering[Rate], Ordering[Ratio] for scalatest matchers
- Change computeAiCapex/computeHybridCapex/computeDigiInvestCost to return PLN
- Refactor decideHybrid/decideTraditional: PLN-typed cost variables, typed comparisons
- Refactor execute: tech imports via PLN * Ratio (no unwrap/rewrap)
- Refactor computePnL: typed monetary flow, no PLN(...) wrappers at end
- Clean up applyInvestment/applyGreenInvestment/applyFdiFlows/applyInformalCitEvasion
- Update 4 test files for PLN-returning capex functions

* Change effectiveWageMult to return Ratio, clean up LaborUnionSpec

* Change computeCapacity to return PLN, update all call sites

* Type process params: lendRate→Rate, bankCanLend→PLN, wage→PLN; decompose applyInformalCitEvasion

* Flatten nested post-processing pipeline in Firm.process into named steps

* Remove unused RunConfig param from Firm.process

* Eliminate .toDouble from applyFdiFlows where possible

* Add Rate / Rate operator, eliminate .toDouble from applyFdiFlows

* Eliminate .toDouble from applyInvestment via typed depRates/klRatios/adjustSpeed

* Add PLN / Ratio operator, eliminate .toDouble from applyInventory

* Eliminate .toDouble from applyGreenInvestment

* Eliminate .toDouble from computePnL: typed depRates, costReplace, energyCostShares, greenDiscount

* Reduce .toDouble in decideTraditional: Ratio arithmetic before conversion

* Reduce .toDouble in decideHybrid: Ratio arithmetic in probability calc

* Reduce .toDouble in decideTraditional: Ratio arithmetic in failRate and goodEff

+++ src/main/scala/sfc/engine/steps/PriceEquityStep.scala
@@ -310,7 +310,7 @@ object PriceEquityStep:
     )
 
     val firmProfits = living2.kahanSumBy { f =>
-      val rev      = Firm.computeCapacity(f) * in.s4.sectorMults(f.sector.toInt) * newPrice
+      val rev      = Firm.computeCapacity(f).toDouble * in.s4.sectorMults(f.sector.toInt) * newPrice
       val labor    = Firm.workerCount(f) * in.s2.newWage * SectorDefs(f.sector.toInt).wageMultiplier
       val other    = p.firm.otherCosts.toDouble * newPrice
       val aiMaint  = f.tech match
